Implement coach sidebar functionality in the app layout, enhancing user experience for coach-related features. Update the coach chat demo text for clarity. Introduce new imports and logic for managing coach sidebar visibility based on user capabilities.
